Try to find dnadiff and mummerplot binaries, instead of hardcoded paths.

+++ src/pipelines/sanity/success.arabidopsis_thaliana.sh
@@ -2,11 +2,26 @@
 
 name=$1
 
-dnadiff -p $name.col0.dnadiff \
+if [ x$name = x ] ; then
+  echo "usage: $0 assembly-prefix"
+  exit 1
+fi
+
+if [ -e "/work/software/MUMmer3.23/mummerplot" ] ; then  PATH=$PATH:/work/software/MUMmer3.23 ; fi
+
+dnadiff=`which dnadiff`
+mummerplot=`which mummerplot`
+
+if [ ! -e $dnadiff -o ! -e $mummerplot ] ; then
+  echo "Didn't find dnadiff or mummerplot in your path."
+  exit 1
+fi
+
+$dnadiff -p $name.col0.dnadiff \
   /data/references/arabidopsis_thaliana/arabidopsis_thaliana_col0.fasta \
   $name.contigs.fasta
 
-mummerplot --fat -t png -p $name.col0.dnadiff $name.col0.dnadiff.delta
+$mummerplot --fat -t png -p $name.col0.dnadiff $name.col0.dnadiff.delta
 
 rm -f $name.col0.dnadiff.1coords
 rm -f $name.col0.dnadiff.1delta
@@ -22,11 +37,11 @@ rm -f $name.col0.dnadiff.gp
 
 
 
-dnadiff -p $name.ler0.dnadiff \
+$dnadiff -p $name.ler0.dnadiff \
   /data/references/arabidopsis_thaliana/arabidopsis_thaliana_ler0.v7.fasta \
   $name.contigs.fasta
 
-mummerplot --fat -t png -p $name.ler0.dnadiff $name.ler0.dnadiff.delta
+$mummerplot --fat -t png -p $name.ler0.dnadiff $name.ler0.dnadiff.delta
 
 rm -f $name.ler0.dnadiff.1coords
 rm -f $name.ler0.dnadiff.1delta
