+++ src/pipelines/sanity/success.bibersteinia_trehalosi.sh
@@ -2,11 +2,26 @@
 
 name=$1
 
-dnadiff -p $name.dnadiff \
+if [ x$name = x ] ; then
+  echo "usage: $0 assembly-prefix"
+  exit 1
+fi
+
+if [ -e "/work/software/MUMmer3.23/mummerplot" ] ; then  PATH=$PATH:/work/software/MUMmer3.23 ; fi
+
+dnadiff=`which dnadiff`
+mummerplot=`which mummerplot`
+
+if [ ! -e $dnadiff -o ! -e $mummerplot ] ; then
+  echo "Didn't find dnadiff or mummerplot in your path."
+  exit 1
+fi
+
+$dnadiff -p $name.dnadiff \
   /data/references/bibersteinia_trehalosi/bibersteinia_trehalosi_USDA-ARS-USMARC-192-CP003745.1.fasta \
   $name.contigs.fasta
 
-mummerplot --fat -t png -p $name.dnadiff $name.dnadiff.delta
+$mummerplot --fat -t png -p $name.dnadiff $name.dnadiff.delta
 
 rm -f $name.dnadiff.1coords
 rm -f $name.dnadiff.1delta
