+++ src/pipelines/sanity/success.escherichia_coli_ne92.sh
@@ -2,11 +2,26 @@
 
 name=$1
 
-/work/software/MUMmer3.23/dnadiff -p $name.dnadiff \
+if [ x$name = x ] ; then
+  echo "usage: $0 assembly-prefix"
+  exit 1
+fi
+
+if [ -e "/work/software/MUMmer3.23/mummerplot" ] ; then  PATH=$PATH:/work/software/MUMmer3.23 ; fi
+
+dnadiff=`which dnadiff`
+mummerplot=`which mummerplot`
+
+if [ ! -e $dnadiff -o ! -e $mummerplot ] ; then
+  echo "Didn't find dnadiff or mummerplot in your path."
+  exit 1
+fi
+
+$dnadiff -p $name.dnadiff \
   /data/references/escherichia_coli_k12_substr__MG1655_uid57779__NC_000913.3.fasta \
   $name.contigs.fasta
 
-/work/software/MUMmer3.23/mummerplot --fat -t png -p $name.dnadiff $name.dnadiff.delta
+$mummerplot --fat -t png -p $name.dnadiff $name.dnadiff.delta
 
 rm -f $name.dnadiff.1coords
 rm -f $name.dnadiff.1delta
